1. tsconfig - no any and no implicit any were added 2. Removed console.logs 3. Added a map.ts file and moved map functions to it 4. Created a types.ts file and moved all interfaces there 5. Changed map UI using correct Tailwind
